Enhance integration test w/ commit with content > 1MB
Test currently fails.  If I try to manually request the content for the commit, it gives: This API returns blobs up to 1 MB in size. The requested blob is too large to fetch via the API, but you can use the Git Data API to request blobs up to 100 MB in size.