Feature/large push limits (#25)
* Fixes #24 Large PushEvents blow up KeyNuker 🔥

https://github.com/tleyden/keynuker/issues/24

* Use separate keypair for integration tests

Too confusing to depend on ordering

* Fix crash

* Detect if existing triggers/rules exist

* Enhance integration test w/ commit with content > 1MB

Test currently fails.  If I try to manually request the content for the commit, it gives: This API returns blobs up to 1 MB in size. The requested blob is too large to fetch via the API, but you can use the Git Data API to request blobs up to 100 MB in size.

* Don’t push large commit by default yet, test still fails with that set to true

* Debug WaitForPushEvent